chore: clean up, small refactoring, updated test samples
