+++ src/main.py
@@ -9,27 +9,31 @@ def main():
     if run_mode == 'train':
         from src.train import start_training
         from src.test.test_model import run_test_training
+        
         start_training()
 
         log("Training completed.")
 
-        log("Starting evaluation...")
-
         run_test_training(mode=os.getenv("TEST_MODE","force_think"))
 
-        log("Evaluation completed. Finish the program.")
         return
     elif run_mode == "test-training":
         from src.test.test_model import run_test_training
+       
         run_test_training(mode=os.getenv("TEST_MODE","force_think"))
+        
         return
     elif run_mode == "test-merging":
         from src.test.test_model import run_test_merging
+        
         run_test_merging(mode=os.getenv("TEST_MODE","force_think"))
+        
         return
     elif run_mode == "test-gguf":
         from src.test.test_model import run_test_gguf
+        
         run_test_gguf(mode=os.getenv("TEST_MODE","force_think"))
+        
         return
     else:
         # error
